requirements.txt
gitpython==3.1.*